dccp: sample kernel module, NULL-CCID ("UDP-like")
This implements an experimental CCID which does not do any congestion control,
i.e. a "UDP-like" CCID.

This is an experimental module. It is not meant for actual deployment, but
rather as sample kernel code, providing a worked example for adding a new
CCID module.

Since the CCID identifier '0' is reserved (RFC 4340, table 5), this 'zero' CCID
uses the the first available experimental CCID number, 248, (RFC 4340, 19.5).

Signed-off-by: Gerrit Renker <gerrit@erg.abdn.ac.uk>
Acked-by: Ivo Calado <ivocalado@embedded.ufcg.edu.br>