dccp: Add sysctl to toggle the local support for ECN
This sysctl disables local ECN support for interpreting ECN bits. 

When set to 0, the host behaves as if it were ECN-incapable: ECN bits are
ignored and the mandatory "Change L(ECN Incapable, 1)" (RFC 4340, 12.1) is
sent during feature negotiation to signal to the other endpoint to signal
that ECN bits will not be interpreted.

The behaviour is slightly different from the tcp_ecn sysctl, since DCCP 
connections are ECN-enabled per default.

Setting ECT codepoints in outgoing packets can not be controlled by this sysctl,
but is possible by setting the sysctl to 0 at the remote (receiving) end. This
signals that the remote end can not read ECN bits. Due to RFC 4340, 12.2, all
outgoing packets are then sent as ECN-incapable.

Inter-operability issues: the host does not set a preference for the remote ECN
Incapable feature. Within the existing feature-negotiation infrastructure, this
ensures that all valid values (ECN Incapable = 0 or 1) are acceptable. 

Signed-off-by: Gerrit Renker <gerrit@erg.abdn.ac.uk>